feat: Parse incoming JSON from Hoogle
In the current version, we get the json array from Hoogle as a
Haskell String, then turn it into ByteString and finally use
aeson's decode to turn it into an array of HoogleJsonResponses.
FromJSON had to be instanced manually for it because of the
field names (namely moduleInfo, packageInfo and typeInfo) being
different from actual json keys. The difference was intentional
in order to avoid keyword conflicts.